kin-news/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dependency-injector==4.40.0
Deprecated==1.2.13
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.4
Faker==15.3.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.4
pyaes==1.6.1
pyasn1==0.4.8
pydantic==1.10.2
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
redis==4.3.4
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.3
Telethon==1.25.4
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
isort==5.11.2
mypy==0.991
mypy-extensions==0.4.3
flake8==6.0.0
